[FEAT] DatabaseRequest class User classes
