--- src/main/java/database/users/UserRequest.java
@@ -1,47 +1,0 @@
-package database.users;
-
-import com.mongodb.DBCollection;
-import com.mongodb.DBObject;
-import com.mongodb.MongoClient;
-import com.mongodb.MongoClientURI;
-import com.mongodb.client.MongoCollection;
-import com.mongodb.client.MongoDatabase;
-import org.bson.Document;
-
-import java.util.function.Consumer;
-
-public class UserRequest {
-
-    protected MongoDatabase database;
-
-
-    public UserRequest() {
-        MongoClientURI uri = new MongoClientURI(
-                "mongodb://udUser:EFBDrRdeE8QAG1lU@cluster0-shard-00-00.wcngd.mongodb.net:27017,cluster0-shard-00-01.wcngd.mongodb.net:27017,cluster0-shard-00-02.wcngd.mongodb.net:27017/<dbname>?ssl=true&replicaSet=atlas-ofie07-shard-0&authSource=admin&retryWrites=true&w=majority");
-        MongoClient mongoClient = new MongoClient(uri);
-        this.database = mongoClient.getDatabase("UndercookedDB");
-    }
-
-    public boolean addUser(String userName, String password) {
-        try {
-            Document newUser = new Document("user_name", userName).append("password", password).append("access", "user");
-            database.getCollection("Users").insertOne(newUser);
-            return true;
-        } catch (Exception exception) {
-            return false;
-        }
-    }
-
-    public boolean getUsers() {
-        return true;
-    }
-
-    public MongoDatabase getDatabase() {
-        return this.database;
-    }
-
-
-    public static void main(String[] args) {
-        UserRequest userRequest = new UserRequest();
-    }
-}
