+++ src/main/java/user/ConnectedUser.java
@@ -1,0 +1,10 @@
+package user;
+
+public class ConnectedUser extends User{
+
+    private String userName;
+
+    public ConnectedUser(String userName) {
+        this.userName = userName;
+    }
+}
